requirements.txt
ollama>=0.1.0
langchain>=0.1.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.30.0
psutil>=5.9.0
pyyaml>=6.0.0
plotly>=5.15.0
langchain-community